backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
twilio==8.10.0
python-decouple==3.8
django-celery-beat==2.5.0
